add brotli submodule and configure multibuild for Travis/Appveyor CI
https://github.com/matthew-brett/multibuild

Multibuild is a set of shell utilities for building Python wheels on travis-ci for OSX and Linux, and on Appveyor for Windows,

The python versions currently enabled are: 2.7, 3.4, 3.5 and 3.6.

Deployment to Github Releases is temporarily configured for the anthrotype/brotli-wheels fork. It will be changed to point to the the upstream repository once the PR is merged.

The PyPI deployment will be configured later.